config

+++ main.py
@@ -62,7 +62,7 @@ def train(**kwargs):
         torch.cuda.set_device(opt.gpu_id)
 
     path_checkpoint = '/content/drive/MyDrive/checkpoints/' + opt.model + '_' + opt.dataset + str(
-        opt.lambda1 * 100) + '.pth'
+        int(opt.lambda1 * 100)) + '.pth'
     if os.path.exists(path_checkpoint):
         print('loading exist model......................')
         checkpoint = torch.load(path_checkpoint)
